fixed : compiled template ownership changed by cron script issue. Our web application needs the compiled templates to be owned by www-data user so that they can be opened and read by the web process (that is run as www-data)
However our cron script (that runs as root) can be the owner of such
compiled templates if cron script accesses those templates before the
web process had a chance to do so. This will result in web process
unable to load such templates.

The fix is to set the ownership back to www-data by the cron worker
at the end of its run.